refactor tests of external tools into /external
