+++ app/src/androidTest/java/com/github/fribourgsdp/radio/external/MusixmatchLyricsGetterTest.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.external
 
 import com.github.fribourgsdp.radio.external.musixmatch.MusixmatchLyricsGetter
 import com.github.fribourgsdp.radio.util.JSONParser
