+++ app/src/androidTest/java/com/github/fribourgsdp/radio/external/MyIRtcEngineEventHandlerTest.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.external
 
 import android.content.Context
 import android.content.Intent
@@ -8,6 +8,7 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.matcher.ViewMatchers
+import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.mockimplementations.MockGameActivity
 import com.github.fribourgsdp.radio.voip.MyIRtcEngineEventHandler
 import org.junit.After
