+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockGameActivity.kt
@@ -1,6 +1,7 @@
 package com.github.fribourgsdp.radio.mockimplementations
 
-import com.github.fribourgsdp.radio.*
+import com.github.fribourgsdp.radio.external.makeMockIRtcEngineEventHandler
+import com.github.fribourgsdp.radio.external.makeMockRtcEngine
 import com.github.fribourgsdp.radio.game.GameActivity
 import com.github.fribourgsdp.radio.voip.VoiceIpEngineDecorator
